requirements.txt
yt-dlp>=2023.12.30
openai-whisper>=20231117
google-generativeai>=0.8.0
spacy>=3.7.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
fastapi>=0.109.0
uvicorn>=0.27.0
